Send JSON-formatted output as the body of the response
This change refactors the code further so that it sends well-formed JSON
as the body of the response, regardless of whether the response was a
success or error. Along with that, the response's content type is set to
application/json, to ensure that the client interprets the response body
properly.